Expose a few more details of underlying backends
This commit updates the various backends to export some more information, where
all of these changes were motivated in getting the tokio-tls test suite running
again. (it's just a bunch of smoke tests really). A summary of the changes are:

* Add an `ErrorExt` trait to each backend to lower a `native_tls::Error` to the
  backend-specific error. This allows inspection of errors codes and such if
  necessary.

* Add `TlsConnectorBuilderExt` to the `security_framework` backend, with one
  function called `anchor_certificates`. This function transitively calls the
  security framework crate's `anchor_certificates` function to set a custom
  trusted certificate for a connection. Along the way this required using the
  `ClientBuilder` type which in turn required changes to work with async I/O.

The changes to the security framework side of things forces this crate to depend
on the changes in sfackler/rust-security-framework#23, so it shouldn't be merged
before that's updated and released!