Use a temp keychain on OSX again
PR #18 undid the original code that looked a bit like this, but it
turned out that the identity would load in to the system keychain on
newer macOS releases, and not work at all on older ones. We now
additionally turn off lock timeouts on the keychain to avoid issues
there with long running processes.